Attempted to create new folder structure with Blueprints but getting errors
