+++ uwsgi.ini
@@ -1,4 +1,3 @@
 [uwsgi]
 module = main
 callable = app
-py-autoreload = 1
